+++ app/Services/Models/Railway/Core/RailwayAchievementAction.php
@@ -32,7 +32,6 @@ class RailwayAchievementAction
     {
         $user = User::find($event->user->id);
         $this->achievement->unlockActionFor($user, 'welcome');
-        $this->notifyAchievementUnlock($user);
     }
 
     public function unDebutATous($event): void
@@ -40,7 +39,6 @@ class RailwayAchievementAction
         $user = User::find($event->user_id);
         if ($user->userRailwayHub()->count() == 1) {
             $this->achievement->unlockActionFor($user, 'un-debut-a-tous', 1);
-            $this->notifyAchievementUnlock($user);
         }
     }
 
@@ -52,7 +50,6 @@ class RailwayAchievementAction
 
         if ($amount >= $this->achievement->goal) {
             $this->achievement->unlockActionFor($user, 'entrepreneur', 1500000);
-            $this->notifyAchievementUnlock($user);
         }
     }
 
@@ -64,7 +61,6 @@ class RailwayAchievementAction
 
         if ($amount >= $this->achievement->goal) {
             $this->achievement->unlockActionFor($user, 'je-rentabilise-ma-societe', 3000000);
-            $this->notifyAchievementUnlock($user);
         }
     }
 
@@ -76,19 +72,8 @@ class RailwayAchievementAction
 
         if ($amount >= $this->achievement->goal) {
             $this->achievement->unlockActionFor($user, 'magnat-ferroviaire', 10000000);
-            $this->notifyAchievementUnlock($user);
         }
     }
 
-    public function notifyAchievementUnlock(User $user): void
-    {
-        if ($this->achievement) {
-            $user->notify(new SendMessageAdminNotification(
-                title: 'Nouveau succès débloquer !',
-                sector: 'alert',
-                type: 'success',
-                message: 'Un nouveau succès à été débloquer !'
-            ));
-        }
-    }
+
 }
